Fixed some warnings

+++ core/algorithms/ImagingAlgorithms/src/ReferenceImageCorrection.cpp
@@ -1424,7 +1424,7 @@ float * ReferenceImageCorrection::SolveLinearEquation(std::map<std::pair<int, in
 
 
     it = values.begin();
-  for (int i=0; i<values.size(); i++){
+  for (int i=0; i<static_cast<int>(values.size()); i++){
 
       if (a && b && c && d && e && f) {
 //            I_int[i] = static_cast<float>(param[0]) + static_cast<float>(param[1])*static_cast<float>(it->first.first)+static_cast<float>(param[2])*static_cast<float>(it->first.first*it->first.first)+static_cast<float>(param[3])*static_cast<float>(it->first.first)*static_cast<float>(it->first.second)+static_cast<float>(param[4])*static_cast<float>(it->first.second)+static_cast<float>(param[5])*static_cast<float>(it->first.second*it->first.second);
